Fix type compatibility compilation issues

+++ ledger/narwhal/batch-certificate/src/lib.rs
@@ -54,7 +54,7 @@ pub enum BatchCertificate<N: Network> {
 
 impl<N: Network> BatchCertificate<N> {
     /// The maximum number of signatures in a batch certificate.
-    pub const MAX_SIGNATURES: usize = BatchHeader::<N>::MAX_CERTIFICATES;
+    pub const MAX_SIGNATURES: u16 = BatchHeader::<N>::MAX_CERTIFICATES;
 }
 
 impl<N: Network> BatchCertificate<N> {
@@ -84,7 +84,7 @@ impl<N: Network> BatchCertificate<N> {
             N::hash_bhp1024(&preimage.to_bits_le())
         }
         // Ensure that the number of signatures is within bounds.
-        ensure!(signatures.len() <= Self::MAX_SIGNATURES, "Invalid number of signatures");
+        ensure!(signatures.len() <= Self::MAX_SIGNATURES as usize, "Invalid number of signatures");
         // Compute the certificate ID.
         if certificate_id != compute_certificate_id(batch_header.batch_id(), &signatures)? {
             bail!("Invalid batch certificate ID")
@@ -103,7 +103,7 @@ impl<N: Network> BatchCertificate<N> {
     /// Initializes a new batch certificate.
     pub fn from(batch_header: BatchHeader<N>, signatures: IndexSet<Signature<N>>) -> Result<Self> {
         // Ensure that the number of signatures is within bounds.
-        ensure!(signatures.len() <= Self::MAX_SIGNATURES, "Invalid number of signatures");
+        ensure!(signatures.len() <= Self::MAX_SIGNATURES as usize, "Invalid number of signatures");
 
         // Verify the signatures are valid.
         cfg_iter!(signatures).try_for_each(|signature| {
